misc/projects/story-to-video/app/server/pyproject.toml
[tool.poetry]
name = "story-to-video-server"
version = "0.1.0"
description = "FastAPI server for story-to-video generation"
authors = ["User"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
fastapi = "^0.114.0"            # upgraded to be compatible with anyio 4.x
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pydantic = "^2.0.0"
python-dotenv = "^1.1.1"
openai = "^1.102.0"
mutagen = "^1.47.0"
google-generativeai = "^0.8.5"
pillow = "^11.3.0"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.28.1"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
